Greedy almost working

+++ SDL_Pathfinding/Node.cpp
@@ -1,7 +1,7 @@
 #include "Node.h"
 #include <iostream>
 
-Node::Node(Vector2D pos, int _weight)
+Node::Node(Vector2D pos, float _weight)
 {
 	SetWeight(_weight);
 	SetPos(pos);
